requirements.txt
# ============================================
# Python Dependencies for LLM Expense Tracker
# ============================================
# CHECKPOINT_1_PROJECT_SETUP

# ============================================
# Core Framework
# ============================================
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.2
python-multipart==0.0.12

# ============================================
# Database (Supabase)
# ============================================
supabase==2.9.0
postgrest==0.16.11

# ============================================
# LLM Integration
# ============================================
# Ollama client
httpx==0.27.2
ollama==0.3.3

# Alternative: Groq (uncomment if using)
# groq==0.11.0

# Alternative: OpenAI (uncomment if using)
# openai==1.54.0

# For structured outputs
jsonschema==4.23.0
pydantic==2.9.0

# ============================================
# Voice Input (Whisper)
# ============================================
openai-whisper==20240930
# Faster implementation (optional)
# faster-whisper==1.0.3

# Audio processing
sounddevice==0.5.1
soundfile==0.12.1
numpy==1.26.4
torch==2.5.1
torchaudio==2.5.1

# ============================================
# HTTP Clients
# ============================================
httpx==0.27.2
requests==2.32.3
aiohttp==3.10.10

# ============================================
# Authentication & Security
# ============================================
PyJWT==2.8.0
python-dotenv==1.0.1

# ============================================
# Utilities
# ============================================
python-dateutil==2.9.0
pytz==2024.2

# ============================================
# Logging & Monitoring
# ============================================
colorlog==6.9.0
structlog==24.4.0

# ============================================
# Testing
# ============================================
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2  # For testing async endpoints

# ============================================
# Development Tools
# ============================================
black==24.10.0
flake8==7.1.1
mypy==1.13.0
isort==5.13.2

# ============================================
# Data Validation
# ============================================
email-validator==2.2.0
phonenumbers==8.13.47

# ============================================
# Caching (optional)
# ============================================
redis==5.2.0
hiredis==3.0.0

# ============================================
# Notes:
# ============================================
# All services are free-tier compatible:
# - Supabase: 500MB database, 50MB storage free
# - Ollama: Completely free, runs locally
# - Groq: 14,400 requests/day free
# - OpenAI: Pay-as-you-go (GPT-4o-mini is cheap)
# - Whisper: Open source, runs locally
# - Teleport API: Completely free

# Installation:
# pip install -r requirements.txt

# For CPU-only (smaller download):
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
